requirements.txt
pandas
numpy 
pyarrow
Flask
gunicorn
pyvirtualdisplay
flask-cors
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
distro==1.9.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
MouseInfo==0.1.3
openai==1.51.2
outcome==1.3.0.post0
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytweening==1.2.0
requests==2.32.3
selenium
setuptools==75.2.0
sniffio==1.3.1
sortedcontainers==2.4.0
tqdm==4.66.5
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
undetected-chromedriver==3.5.5
urllib3==2.2.3
websocket-client==1.8.0
websockets==13.1
wsproto==1.2.0
webdriver-manager
google-generativeai
python-dotenv
pdf2image
azure-data-tables
azure-storage-blob
beautifulsoup4 
flask-socketio 
eventlet
streamlit
azure-data-tables
PyPDF2
fpdf
pdfminer-six
python-docx
Jinja2
docx2txt
streamlit-ext==0.1.9
